Trace.Forward.Configuration.TraceObject

Documentation

data AcceptorConfiguration lo

Acceptor configuration, parameterized by trace item's type.

Constructors

AcceptorConfiguration

Fields

acceptorTracer :: !(Tracer IO (TraceSendRecv (TraceObjectForward lo)))

The tracer that will be used by the acceptor in its network layer.

whatToRequest :: !NumberOfTraceObjects

The request specifies how many TraceObjects will be requested.

shouldWeStop :: !(TVar Bool)

TVar that can be used as a brake: if an external thread sets it to True, the acceptor will send MsgDone message to the forwarder and their session will be closed.

data ForwarderConfiguration lo

Forwarder configuration, parameterized by trace item's type.

Constructors

ForwarderConfiguration

Fields

forwarderTracer :: !(Tracer IO (TraceSendRecv (TraceObjectForward lo)))

The tracer that will be used by the forwarder in its network layer.

queueSize :: !Word

The size of the internal queue for tracing items. Use a size suitable for the beginning of the session, to avoid queue overflows, because initially there is no connection with acceptor yet, and the number of tracing items after the node starts may be very big. At the same time choose a number that reduces memory usage in the node.