Trace.Forward.Forwarding

Documentation

data InitForwardingConfig

Config record to initialise trace forwarding

Constructors

InitForwardingNone
Only construct relevant values, but do not actually run forwarding

InitForwardingWith
Run forwarding with the provided settings

Fields

initNetworkMagic :: !NetworkMagic

Forwarding is always tied to a singular networkId

initEKGStore :: !(Maybe Store)

A metrics store to be forwarded (optional)

initHowToConnect :: !HowToConnect

A LocalPipe or RemoteSocket

initForwarderMode :: !ForwarderMode

Run as Initiator or Responder

initOnForwardInterruption :: !(Maybe (SomeException -> IO ()))

Optional handler when forwarding connection is interrupted (may be temporary or permanent) default: no action

initOnQueueOverflow :: !(Maybe ([TraceObject] -> IO ()))

Optional handler when forwarding queue overflows (argument are objects dropped from queue) default: print one-liner to stderr, indicating object count and timestamps of first and last object

initForwarding :: forall m. MonadIO m => IOManager -> TraceOptionForwarder -> InitForwardingConfig -> m (ForwardSink TraceObject, DataPointStore)

initForwardingDelayed :: forall m. MonadIO m => IOManager -> TraceOptionForwarder -> InitForwardingConfig -> m (ForwardSink TraceObject, DataPointStore, IO ())