Trace.Forward.Protocol.DataPoint.Acceptor

Description

A view of the trace forwarding/accepting protocol from the point of view of the client.

For execution, a conversion into the typed protocol is provided.

Documentation

data DataPointAcceptor m a where

Constructors

SendMsgDataPointsRequest :: [DataPointName] -> (DataPointValues -> m (DataPointAcceptor m a)) -> DataPointAcceptor m a

SendMsgDone :: m a -> DataPointAcceptor m a

dataPointAcceptorPeer :: Monad m => DataPointAcceptor m a -> Client DataPointForward 'NonPipelined 'StIdle m a

Interpret a particular action sequence into the client side of the protocol.