Trace.Forward.Protocol.DataPoint.Forwarder

Documentation

data DataPointForwarder m a

Constructors

DataPointForwarder

Fields

recvMsgDataPointsRequest :: [DataPointName] -> m DataPointValues

The acceptor sent us a request for new DataPoints.

recvMsgDone :: m a

The acceptor terminated. Here we have a pure return value, but we could have done another action in m if we wanted to.

dataPointForwarderPeer :: Monad m => DataPointForwarder m a -> Server DataPointForward 'NonPipelined 'StIdle m a

Interpret a particular action sequence into the server side of the protocol.