Trace.Forward.Protocol.TraceObject.Acceptor

Description

A view of the trace forwarding/accepting protocol from the point of view of the client.

For execution, a conversion into the typed protocol is provided.

Documentation

data TraceObjectAcceptor lo m a where

Constructors

SendMsgTraceObjectsRequest :: TokBlockingStyle blocking -> NumberOfTraceObjects -> (BlockingReplyList blocking lo -> m (TraceObjectAcceptor lo m a)) -> TraceObjectAcceptor lo m a

SendMsgDone :: m a -> TraceObjectAcceptor lo m a

traceObjectAcceptorPeer :: Monad m => TraceObjectAcceptor lo m a -> Client (TraceObjectForward lo) 'NonPipelined 'StIdle m a

Interpret a particular action sequence into the client side of the protocol.