Trace.Forward.Protocol.TraceObject.Forwarder

Documentation

data TraceObjectForwarder lo m a

Constructors

TraceObjectForwarder

Fields

recvMsgTraceObjectsRequest :: forall blocking. TokBlockingStyle blocking -> NumberOfTraceObjects -> m (BlockingReplyList blocking lo)

The acceptor sent us a request for new TraceObjects.

recvMsgDone :: m a

The acceptor terminated. Here we have a pure return value, but we could have done another action in m if we wanted to.

traceObjectForwarderPeer :: forall m lo a. Monad m => TraceObjectForwarder lo m a -> Server (TraceObjectForward lo) 'NonPipelined 'StIdle m a

Interpret a particular action sequence into the server side of the protocol.