Trace.Forward.Utils.DataPoint

Documentation

data DataPoint where

Type wrapper for some value of type v. The only reason we need this wrapper is an ability to store different values in the same DataPointStore.

Please note that when the acceptor application will read the value of type v from the store, this value is just as unstructured JSON, but not Haskell value of type v. That's why FromJSON instance for type v should be available for the acceptor application, to decode unstructured JSON.

Constructors

DataPoint :: forall v. (ToJSON v, NFData v) => !v -> DataPoint

type DataPointStore = TVar (Map DataPointName DataPoint)

data DataPointRequestor

Since DataPointForward protocol does not assume the stream of requests/replies, we use the TVars to provide to acceptor's side an ability to ask DataPoints explicitly.

Constructors

DataPointRequestor

Fields

askDataPoints :: !(TVar Bool)

The "ask flag": we use it to notify that we want DataPoints.

dataPointsNames :: !(TVar [DataPointName])

The names of DataPoints we need.

dataPointsReply :: !(TMVar DataPointValues)

The list of received DataPoint's values. By default it's empty, but when DataPoints are received they will be stored here.

initDataPointStore :: IO DataPointStore

initDataPointRequestor :: IO DataPointRequestor

writeToStore :: DataPointStore -> DataPointName -> DataPoint -> IO ()

Write DataPoint to the store.

readFromStore :: DataPointStore -> DataPointForwarder IO ()

Read DataPoints from the store. Please note that we don't care what's inside of DataPoint, we just know it can be encoded to JSON.

askForDataPoints :: DataPointRequestor -> [DataPointName] -> IO DataPointValues