Trace.Forward.Protocol.DataPoint.Type

Description

The type of the DataPoint forwarding/accepting protocol.

Documentation

type DataPointName = Text

A kind to identify our protocol, and the types of the states in the state transition diagram of the protocol.

IMPORTANT NOTE: the following terminology is used:

From the protocol's point of view, two peers talk to each other: the forwarder and the acceptor.

The forwarder is an application that collects DataPoints and sends them to the acceptor by request (with MsgDataPointsReply).

The acceptor is an application that receives DataPoints from the forwarder.

You can think of the acceptor as a client, and the forwarder as a server. After the connection is established, the acceptor asks for DataPoints, the forwarder replies to it.

type DataPointValue = ByteString

type DataPointValues = [(DataPointName, Maybe DataPointValue)]

data DataPointForward where

Constructors

StIdle :: DataPointForward
Both acceptor and forwarder are in idle state. The acceptor can send a request for a list of DataPoints (MsgDataPointsRequest); the forwarder is waiting for a request, it will replay with MsgDataPointsReply.

StBusy :: DataPointForward
The acceptor has sent a next request for DataPoints. The acceptor is now waiting for a reply, and the forwarder is busy getting ready to send a reply with new list of DataPoints.

StDone :: DataPointForward
Both the acceptor and forwarder are in the terminal state. They're done.

data family Message ps (st :: ps) (st' :: ps)

The messages for this protocol. It is expected to be a GADT that is indexed by the from and to protocol states. That is the protocol state the message transitions from, and the protocol state it transitions into. These are the edges of the protocol state transition system.

data family ClientHasAgency (st :: ps)

Tokens for those protocol states in which the client has agency.

data family ServerHasAgency (st :: ps)

Tokens for those protocol states in which the server has agency.

data family NobodyHasAgency (st :: ps)

Tokens for terminal protocol states in which neither the client nor server has agency.