Trace.Forward.Protocol.TraceObject.Type

Description

The type of the trace forwarding/accepting protocol.

Documentation

data TraceObjectForward lo where

Constructors

StIdle :: TraceObjectForward lo
Both acceptor and forwarder are in idle state. The acceptor can send a request for a list of TraceObjects (MsgTraceObjectsRequest); the forwarder is waiting for a request. It will replay with MsgTraceObjectsReply.

Node's info is an important information about the node, such as its protocol, version, start time, etc. It is assuming that the node must provide this information.

StBusy :: StBlockingStyle -> TraceObjectForward lo
The acceptor has sent a next request for TraceObjects. The acceptor is now waiting for a reply, and the forwarder is busy getting ready to send a reply with new list of TraceObjects.

There are two sub-states for this, for blocking and non-blocking cases.

StDone :: TraceObjectForward lo
Both the acceptor and forwarder are in the terminal state. They're done.

data TokBlockingStyle (k :: StBlockingStyle) where

The value level equivalent of StBlockingStyle.

This is also used in MsgTraceObjectsRequest where it is interpreted (and can be encoded) as a Bool with True for blocking, and False for non-blocking.

Constructors

TokBlocking :: TokBlockingStyle 'StBlocking

TokNonBlocking :: TokBlockingStyle 'StNonBlocking

data family Message ps (st :: ps) (st' :: ps)

The messages for this protocol. It is expected to be a GADT that is indexed by the from and to protocol states. That is the protocol state the message transitions from, and the protocol state it transitions into. These are the edges of the protocol state transition system.

data family ClientHasAgency (st :: ps)

Tokens for those protocol states in which the client has agency.

data family ServerHasAgency (st :: ps)

Tokens for those protocol states in which the server has agency.

data family NobodyHasAgency (st :: ps)

Tokens for terminal protocol states in which neither the client nor server has agency.

newtype NumberOfTraceObjects

A kind to identify our protocol, and the types of the states in the state transition diagram of the protocol.

IMPORTANT NOTE: the following terminology is used:

From the protocol's point of view, two peers talk to each other: the forwarder and the acceptor.

The forwarder is an application that collects TraceObjects and sends them to the acceptor by request (with MsgTraceObjectsReply).

The acceptor is an application that receives TraceObjects from the forwarder.

You can think of the acceptor as a client, and the forwarder as a server. After the connection is established, the acceptor asks for TraceObjects, the forwarder replies to it.

The acceptor will send this request to the forwarder.

Constructors

NumberOfTraceObjects

Fields

nTraceObjects :: Word16

data BlockingReplyList (blocking :: StBlockingStyle) lo where

We have requests for lists of things. In the blocking case the corresponding reply must be non-empty, whereas in the non-blocking case an empty reply is fine.

Constructors

BlockingReply :: NonEmpty lo -> BlockingReplyList 'StBlocking lo

NonBlockingReply :: [lo] -> BlockingReplyList 'StNonBlocking lo